fix: address round-1 review — a zero-norm row is not an exact-zero diagonal
The self path's move to the strict upper triangle was documented as a
last-bit reassociation because "the discarded diagonal is exactly 0 for
both gemm metrics". That is false on one kind of row, and the claim was
repeated in docs/scanpy.md, the PR body, and a test comment asserting
the boundary was uncoverable.

`normalize_rows` leaves a zero-norm row as zeros, so its cosine Gram
diagonal is `1 - 0 = 1`, not `1 - 1 = 0`. Two all-zero rows: the full
square gives 1.0, the triangle 0.5. Measured, not reasoned.

Which way that cuts is worth stating, because it is the opposite of a
regression: the triangle is what `backend="scalar"` has always returned
and what `sklearn.metrics.pairwise.cosine_distances` returns, since it
forces the self diagonal to 0 when `X is Y`. So the `auto` default and
`scalar` used to *disagree* on such input and now agree.

  - `gemm_self_skips_the_diagonal_of_a_zero_norm_row` pins it, with a
    premise assertion that the full-square convention really differs on
    the fixture. It also closes the `j_start = t` vs `t + 1` boundary
    that the old comment called untestable — the mutation reddens now.
  - `compute_energy_distance_gpu` still called the full-square kernel for
    its self terms, so CPU and GPU would have diverged. Its self calls go
    through a correction that subtracts the zero-norm diagonal. It is
    exactly 0 for euclidean and for any input without a zero-norm row, so
    every existing GPU fixture is unchanged — and both halves are
    unit-tested on the CPU, one against the CPU kernels themselves.

Also from the same round:

  - The budget contract overstated what it covers, found independently by
    two reviewers. Cosine's `normalize_rows` copies (~800 MB self,
    ~1.6 GB cross at 100K x 2000 f32) sit outside it, and the block is
    `max(budget, one Gram row)` because the planner will not return a
    zero-row block. Restated precisely in distances.rs, scanpy.md and
    architecture.md.
  - "The knob never changes what is summed" contradicted the kernel's own
    doc. Same pairs and same reduction order; a different block width
    repanels the gemm and moves Gram ulps. Only one block is bit-identical.
  - The new performance.md numbers now carry the same manifest
    disclosure the covariance-PCA microbenchmark section uses, rather
    than silently not meeting docs/benchmark_manifest.md.
  - `plan_gram_row_block` clamps in u64 before narrowing, so a 32-bit
    target cannot truncate a large quotient into a tiny block.
  - `gpu_mean_pairwise_distance`'s `is_self` gains the `na == nb` guard
    the CPU side got, so the fix does not rot into a CPU/GPU asymmetry.
  - `row_sq_norms` uses `par_chunks_exact` instead of manual indexing.

+++ scx-accel/src/mem_budget.rs
@@ -80,7 +80,10 @@ pub fn pairwise_memory_budget() -> u64 {
 ///   clamp is what keeps the cross path bit-identical.)
 pub fn plan_gram_row_block(n_a: usize, n_b: usize, elem_bytes: usize, budget: u64) -> usize {
     let per_row = (n_b as u64).saturating_mul(elem_bytes as u64).max(1);
-    ((budget / per_row) as usize).clamp(1, n_a.max(1))
+    // Clamp in `u64` *before* narrowing: on a 32-bit target `budget / per_row`
+    // can exceed `usize::MAX` and a direct cast would truncate — plausibly to a
+    // tiny value or 0, silently shrinking the block instead of widening it.
+    ((budget / per_row).clamp(1, n_a.max(1) as u64)) as usize
 }
 
 /// Pure parse of the `SCX_ACCEL_NUM_THREADS` value: `Some(n)` for a positive
